Relocate photon context-pack truncate + adopted-id prep (Phase 8 batch 4, #688)
Issue #688 (parent #680, Phase 8) batch 4: migrate the photon
context-pack byte-truncation helper, the adopted-id sanitize+cap
helper, the supporting `SanitizedAdoptedIds` struct, and the
`MAX_PHOTON_CONTEXT_PACK_PROMPT_BYTES` const from `turn.rs` to
`photon_feedback_derive.rs`.

Migrated to `photon_feedback_derive.rs`:

* `pub const MAX_PHOTON_CONTEXT_PACK_PROMPT_BYTES: usize = 8192`
* `pub fn truncate_photon_context_pack(s) -> (String, bool)` (UTF-8
    safe byte truncation, re-exported via `loop_run.rs`)
* `pub(crate) struct SanitizedAdoptedIds { list, truncated }`
* `pub(crate) fn prepare_adopted_ids_for_evaluate(raw, shadow_mode)
    -> SanitizedAdoptedIds` (re-runs `sanitize_summary_id` per id,
    caps at `MAX_PHOTON_EVAL_ADOPTED_IDS`, shadow guard short-circuits
    to empty)

`photon_feedback_derive.rs` imports gain `crate::photon` and
`crate::photon::prompt::sanitize_summary_id`.

`turn.rs` adjustments:

* Removed the 2 fn + 1 struct + 1 const definitions (originally at
  lines 483-546).
* Added `prepare_adopted_ids_for_evaluate` to the existing
  `use super::photon_feedback_derive::{...}` block.
* Test mod `prepare_adopted_ids_for_evaluate_tests` rewrote its
  import to `super::super::photon_feedback_derive::*`.

`loop_run.rs` adjustments:

* Consolidated the 2 separate `pub use` lines into a single
  `pub use photon_feedback_derive::{
    MAX_PHOTON_CONTEXT_PACK_PROMPT_BYTES, build_photon_injection_message,
    truncate_photon_context_pack,
  };` block. `tests/photon_turn_hook_smoke.rs` import paths
  (`anvil::agent::loop_run::*`) are unchanged.

LOC delta:

* `turn.rs`: 29,443 → 29,378 (-65 LOC)
* `photon_feedback_derive.rs`: 485 → 552 (+67 LOC)
* Cumulative `turn.rs`: 39,489 → 29,378 (-10,111 LOC, -25.6%)

`pub(crate)` for `SanitizedAdoptedIds` / `prepare_adopted_ids_for_evaluate`,
`pub` for `MAX_PHOTON_CONTEXT_PACK_PROMPT_BYTES` /
`truncate_photon_context_pack` (matching pre-extraction visibility).

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
